Move db component to ui folder
