+++ site/gatsby-site/src/components/forms/QuickAddForm.js
@@ -4,7 +4,7 @@ import { useFormik } from 'formik';
 import * as Yup from 'yup';
 
 import Link from 'components/ui/Link';
-import DBConnecting from 'components/DBConnecting';
+import DBConnecting from 'components/ui/DBConnecting';
 import { useUserContext } from 'contexts/userContext';
 import useToastContext, { SEVERITY } from '../../hooks/useToast';
